Separate out validation module
