+++ ftorch_utils/__init__.py
@@ -4,10 +4,18 @@ from .torchscript import (
     script_to_torchscript,
     trace_to_torchscript,
 )
+from .validation import (
+    validate_input_model_file,
+    validate_output_model_file,
+    validate_input_tensor_file,
+)
 
 __all__ = [
     "load_pytorch",
     "load_torchscript",
     "script_to_torchscript",
     "trace_to_torchscript",
+    "validate_input_model_file",
+    "validate_output_model_file",
+    "validate_input_tensor_file",
 ]
